- 코드 정리
- 여러 페이지로 나누여져있었던 3페이지의 통계를 한 페이지로 합침
