+++ src/main/java/com/tech/EPL/realty/controller/ELPRealtyController.java
@@ -26,16 +26,11 @@ public class ELPRealtyController {
 		
 	@GetMapping("/realty2")
 	public String realty2(Model model) {
-		serviceGroup.getAvgData(model);
 		return "epl/realty2";
 	}
 	
 	@GetMapping("/realty3")
 	public String realty3(Model model) {
-		serviceGroup.getAvgData(model);
-//		String fileName = "C:\\team Albamon\\EPL_Seoul 자료\\부동산 실거래가 정보 모음\\월세_실거래가_년도별_엑셀\\파이썬작업끝난곳\\월세_실거래가_종합_문자_py.csv";
-//		String type = "전월세";
-//		serviceGroup.rentFileDBInsert(fileName, type);
 		return "epl/realty3";
 	}
 	
@@ -54,23 +49,9 @@ public class ELPRealtyController {
 		return "epl/realty6";
 	}
 
-	// 파일업로드
-	@PostMapping("/rent-file")
-	public String rentFileData(@RequestParam String fileName,
-			@RequestParam String type) {
-//		serviceGroup.rentFileDBInsert(fileName, type);
-		
-		return "epl/realty";
-	}
 
-	// 파일업로드
-	@PostMapping("/buy-file")
-	public String buyFileData(@RequestParam String fileName,
-			@RequestParam String type) {
-//		serviceGroup.buyFileDBInsert(fileName, type);
-		
-		return "epl/realty";
-	}
+
+
 	
 	
 }
