+++ src/main/java/com/tech/EPL/realty/service/RealtyAvgData.java
@@ -8,11 +8,11 @@ import org.springframework.ui.Model;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tech.EPL.interfaces.ExecutionModel;
-import com.tech.EPL.realty.dto.YearsStatBuying;
-import com.tech.EPL.realty.dto.YearsStatJeonse;
-import com.tech.EPL.realty.dto.YearsStatRent;
 import com.tech.EPL.realty.mapper.RealtyMapper;
 
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @Service
 public class RealtyAvgData implements ExecutionModel {
 	
@@ -27,42 +27,19 @@ public class RealtyAvgData implements ExecutionModel {
 	
 	@Override
 	public void execution(Model model) {
-		ArrayList<YearsStatBuying> buyingStat;
-		ArrayList<YearsStatRent> rentStat;
-		ArrayList<YearsStatJeonse> jeonseStat;
 		
+		addModelMethod(model, "buyingStat", realtyMapper.getYearsStatBuying());
+		addModelMethod(model, "rentStat", realtyMapper.getYearsStatRent());
+		addModelMethod(model, "jeonseStat", realtyMapper.getYearsStatJeonse());
+
+	}
+	
+	private <T> void addModelMethod(Model model, String keyName, ArrayList<T> list) {
 		try {
-			buyingStat = realtyMapper.getYearsStatBuying();
-			rentStat = realtyMapper.getYearsStatRent();
-			jeonseStat = realtyMapper.getYearsStatJeonse();
-		} catch (Exception e) {
-			System.err.println("Database operation failed: " + e.getMessage());
-			throw e;
-		}
-//		for (int i = 0; i < buyingStat.size(); i++) {
-//			System.out.println(
-//					buyingStat.get(i).toString() // + " | " +
-//					rentStat.get(i).toString() + " | " +
-//					jeonseStat.get(i).toString()
-//					
-//					);
-//		}
-		String jsonBuyingStat = "";
-		String jsonRentStat = "";
-		String jsonJeonseStat = "";
-		
-		try {
-			jsonBuyingStat = mapper.writeValueAsString(buyingStat);
-			jsonRentStat = mapper.writeValueAsString(rentStat);
-			jsonJeonseStat = mapper.writeValueAsString(jeonseStat);
+			model.addAttribute(keyName, mapper.writeValueAsString(list));
 		} catch (JsonProcessingException e) {
-			e.printStackTrace();
+			log.error("Json Processing Exception", e);
 		}
-				
-		model.addAttribute("buyingStat", jsonBuyingStat);
-		model.addAttribute("rentStat", jsonRentStat);
-		model.addAttribute("jeonseStat", jsonJeonseStat);
-		
 	}
 	
 }
