+++ src/main/resources/static/js/realty/realty_body5.js
@@ -1,0 +1,28 @@
+/**
+ * 
+ */
+$(document).ready(function() {
+	loadData();
+	
+	
+	function loadData() {
+		fetch('', {
+			method: 'GET',
+			headers: {
+				'Content-Type': 'application/json'
+			}
+		})
+		.then(response => {
+			if (!response.ok) {
+			    throw new Error(`HTTP error. status: ${response.status}`);
+			} 
+			return response.json();
+		})
+		.then(data => {
+			console.log(data);
+		})
+		.catch(error => {
+			console.error(error);
+		});
+	}
+});
