+++ src/main/webapp/WEB-INF/views/body/realty/realty1.jsp
@@ -7,6 +7,7 @@
 <meta charset="UTF-8">
 <script	src="https://cdn.jsdelivr.net/npm/chart.js@3.7.1/dist/chart.min.js"></script>
 <link rel="stylesheet" href="/static/css/tiles.css">
+<link rel="stylesheet" href="/static/css/realty/on_off.css" />
 <title>realty_body</title>
 </head>
 
@@ -14,14 +15,57 @@
 	<script>
 		$(document).ready(function(){
 			let buyingStat = ${buyingStat };
+			let rentStat = ${rentStat };
+			let jeonseStat = ${jeonseStat };
 			
-			chart(buyingStat);
+			chart(buyingStat, rentStat, jeonseStat);
 		});
 	</script>
 
-	<h1>매매 실거래가 통계 </h1>
+	<h1>연도별 실거래가 통계 </h1>
 
-	<canvas id="chart-buying"></canvas>
+	<label for="buying">
+		<input type="radio" name="seType" id="buying" value="buying" checked />
+		매매 평균 
+	</label>
+	<label for="rent">
+		<input type="radio" name="seType" id="rent" value="rent" />
+		월세 평균 
+	</label>
+	<label for="jeonse">
+		<input type="radio" name="seType" id="jeonse" value="jeonse" />
+		전세 보증금 평균  
+	</label>
+	
+	<!--  -->
+	<div id="buying-container" class="on">
+		<canvas id="chart-buying"></canvas>
+	</div>
+	
+	<!--  -->
+	<div id="rent-container" class="off">
+		<label for="rtfe">
+			<input type="radio" name="avg" id="rtfe" value="rtfe" checked />
+			월세 통계 
+		</label>
+		<label for="grfe">
+			<input type="radio" name="avg" id="grfe" value="grfe" />
+			보증금 통계 
+		</label>
+		
+		<div id="rtfe-chart" class="on">
+			<canvas id="chart-rent-rtfe"></canvas>
+		</div>
+		<div id="grfe-chart" class="off">
+			<canvas id="chart-rent-grfe"></canvas>	
+		</div>
+	</div>
+	
+	<!--  -->
+	<div id="jeonse-container" class="off">
+			<canvas id="chart-jeonse"></canvas>	
+	</div>
+	
 	<script src="/static/js/realty/realty_body1.js"></script>
 	<script src="/static/js/realty/realty_navbar.js"></script>
 </body>
