+++ src/main/webapp/WEB-INF/views/body/realty/realty5.jsp
@@ -16,6 +16,7 @@
 <body>
   <h1> realty_body 5 </h1>
 
+<script src="/static/js/realty/realty_body5.js"></script>
 <script src="/static/js/realty/realty_navbar.js"></script>
 </body>
 </html>
